requirements.txt
Django==2.2.6
django-admin-generator==2.2.0
django-admin-interface==0.12.0
django-filter==2.2.0
django-tinymce4-lite==1.7.5
djangorestframework==3.10.3
drf-dynamic-fields==0.3.1
graphene-django==2.7.1
drf_yasg